+++ bin/auto-pr.ps1
@@ -11,6 +11,10 @@
 .PARAMETER App
     Manifest name to search.
     Placeholders are supported.
+.PARAMETER CommitMessageFormat
+    The format of the commit message.
+    <app> will be replaced with the file name of manifest.
+    <version> will be replaced with the version of the latest manifest.
 .PARAMETER Dir
     The directory where to search for manifests.
 .PARAMETER Push
@@ -43,6 +47,7 @@ param(
     [String] $Upstream,
     [String] $OriginBranch = 'master',
     [String] $App = '*',
+    [String] $CommitMessageFormat = '<app>: Update to version <version>',
     [ValidateScript( {
         if (!(Test-Path $_ -Type Container)) {
             throw "$_ is not a directory!"
@@ -61,7 +66,6 @@ param(
 
 . "$PSScriptRoot\..\lib\manifest.ps1"
 . "$PSScriptRoot\..\lib\json.ps1"
-. "$PSScriptRoot\..\lib\unix.ps1"
 
 if ($App -ne '*' -and (Test-Path $App -PathType Leaf)) {
     $Dir = Split-Path $App
@@ -87,7 +91,7 @@ Optional options:
     exit 0
 }
 
-if (is_unix) {
+if ($IsLinux -or $IsMacOS) {
     if (!(which hub)) {
         Write-Host "Please install hub ('brew install hub' or visit: https://hub.github.com/)" -ForegroundColor Yellow
         exit 1
@@ -110,7 +114,7 @@ function execute($cmd) {
     return $output
 }
 
-function pull_requests($json, [String] $app, [String] $upstream, [String] $manifest) {
+function pull_requests($json, [String] $app, [String] $upstream, [String] $manifest, [String] $commitMessage) {
     $version = $json.version
     $homepage = $json.homepage
     $branch = "manifest/$app-$version"
@@ -127,7 +131,7 @@ function pull_requests($json, [String] $app, [String] $upstream, [String] $manif
     Write-Host "Creating update $app ($version) ..." -ForegroundColor DarkCyan
     execute "hub checkout -b $branch"
     execute "hub add $manifest"
-    execute "hub commit -m '${app}: Update to version $version'"
+    execute "hub commit -m '$commitMessage"
     Write-Host "Pushing update $app ($version) ..." -ForegroundColor DarkCyan
     execute "hub push origin $branch"
 
@@ -142,7 +146,7 @@ function pull_requests($json, [String] $app, [String] $upstream, [String] $manif
     Write-Host "hub pull-request -m '<msg>' -b '$upstream' -h '$branch'" -ForegroundColor Green
 
     $msg = @"
-$app`: Update to version $version
+$commitMessage
 
 Hello lovely humans,
 a new version of [$app]($homepage) is available.
@@ -155,7 +159,7 @@ a new version of [$app]($homepage) is available.
     hub pull-request -m "$msg" -b "$upstream" -h "$branch"
     if ($LASTEXITCODE -gt 0) {
         execute 'hub reset'
-        abort "Pull Request failed! (hub pull-request -m '${app}: Update to version $version' -b '$upstream' -h '$branch')"
+        abort "Pull Request failed! (hub pull-request -m '$commitMessage' -b '$upstream' -h '$branch')"
     }
 }
 
@@ -189,7 +193,7 @@ hub diff --name-only | ForEach-Object {
         return
     }
     $version = $json.version
-
+    $CommitMessage = $CommitMessageFormat -replace '<app>',$app -replace '<version>',$version
     if ($Push) {
         Write-Host "Creating update $app ($version) ..." -ForegroundColor DarkCyan
         execute "hub add $manifest"
@@ -198,12 +202,12 @@ hub diff --name-only | ForEach-Object {
         $status = execute 'hub status --porcelain -uno'
         $status = $status | Where-Object { $_ -match "M\s{2}.*$app.json" }
         if ($status -and $status.StartsWith('M  ') -and $status.EndsWith("$app.json")) {
-            execute "hub commit -m '${app}: Update to version $version'"
+            execute "hub commit -m '$commitMessage'"
         } else {
             Write-Host "Skipping $app because only LF/CRLF changes were detected ..." -ForegroundColor Yellow
         }
     } else {
-        pull_requests $json $app $Upstream $manifest
+        pull_requests $json $app $Upstream $manifest $CommitMessage
     }
 }