+++ lib/autoupdate.ps1
@@ -248,7 +248,7 @@ function get_hash_for_app([String] $app, $config, [String] $version, [String] $u
         'sourceforge' {
             # change the URL because downloads.sourceforge.net doesn't have checksums
             $hashfile_url = (strip_filename (strip_fragment "https://sourceforge.net/projects/$($matches['project'])/files/$($matches['file'])")).TrimEnd('/')
-            $hash = find_hash_in_textfile $hashfile_url $substitutions '"$basename":.*?"sha1":\s"([a-fA-F0-9]{40})"'
+            $hash = find_hash_in_textfile $hashfile_url $substitutions '"$basename":.*?"sha1":\s*"([a-fA-F0-9]{40})"'
         }
     }
 
@@ -274,7 +274,7 @@ function get_hash_for_app([String] $app, $config, [String] $version, [String] $u
         return $null
     }
     $file = fullpath (cache_path $app $version $url)
-    $hash = compute_hash $file 'sha256'
+    $hash = (Get-FileHash -Path $file -Algorithm SHA256).Hash.ToLower()
     write-host -f DarkYellow 'Computed hash: ' -NoNewline
     write-host -f Green $hash
     return $hash