+++ libexec/scoop-download.ps1
@@ -85,7 +85,7 @@ foreach ($curr_app in $apps) {
 
     $curr_check_hash = $check_hash
     if ($version -eq 'nightly') {
-        $version = nightly_version $(get-date)
+        $version = nightly_version
         $curr_check_hash = $false
     }