+++ test/Scoop-Core.Tests.ps1
@@ -1,10 +1,8 @@
-. "$PSScriptRoot\Scoop-TestLib.ps1"
-. "$PSScriptRoot\..\lib\core.ps1"
-. "$PSScriptRoot\..\lib\install.ps1"
-. "$PSScriptRoot\..\lib\unix.ps1"
-
-$repo_dir = (Get-Item $MyInvocation.MyCommand.Path).directory.parent.FullName
-$isUnix = is_unix
+BeforeAll {
+    . "$PSScriptRoot\Scoop-TestLib.ps1"
+    . "$PSScriptRoot\..\lib\core.ps1"
+    . "$PSScriptRoot\..\lib\install.ps1"
+}
 
 Describe 'Get-AppFilePath' -Tag 'Scoop' {
     BeforeAll {
@@ -128,15 +126,14 @@ Describe 'is_directory' -Tag 'Scoop' {
     }
 }
 
-Describe 'movedir' -Tag 'Scoop' {
-    $extract_dir = 'subdir'
-    $extract_to = $null
-
+Describe 'movedir' -Tag 'Scoop', 'Windows' {
     BeforeAll {
         $working_dir = setup_working 'movedir'
+        $extract_dir = 'subdir'
+        $extract_to = $null
     }
 
-    It 'moves directories with no spaces in path' -Skip:$isUnix {
+    It 'moves directories with no spaces in path' {
         $dir = "$working_dir\user"
         movedir "$dir\_tmp\$extract_dir" "$dir\$extract_to"
 
@@ -144,7 +141,7 @@ Describe 'movedir' -Tag 'Scoop' {
         "$dir\_tmp\$extract_dir" | Should -Not -Exist
     }
 
-    It 'moves directories with spaces in path' -Skip:$isUnix {
+    It 'moves directories with spaces in path' {
         $dir = "$working_dir\user with space"
         movedir "$dir\_tmp\$extract_dir" "$dir\$extract_to"
 
@@ -157,7 +154,7 @@ Describe 'movedir' -Tag 'Scoop' {
         "$dir\_tmp" | Should -Not -Exist
     }
 
-    It 'moves directories with quotes in path' -Skip:$isUnix {
+    It 'moves directories with quotes in path' {
         $dir = "$working_dir\user with 'quote"
         movedir "$dir\_tmp\$extract_dir" "$dir\$extract_to"
 
@@ -166,14 +163,14 @@ Describe 'movedir' -Tag 'Scoop' {
     }
 }
 
-Describe 'shim' -Tag 'Scoop' {
+Describe 'shim' -Tag 'Scoop', 'Windows' {
     BeforeAll {
         $working_dir = setup_working 'shim'
         $shimdir = shimdir
         $(ensure_in_path $shimdir) | Out-Null
     }
 
-    It "links a file onto the user's path" -Skip:$isUnix {
+    It "links a file onto the user's path" {
         { Get-Command 'shim-test' -ea stop } | Should -Throw
         { Get-Command 'shim-test.ps1' -ea stop } | Should -Throw
         { Get-Command 'shim-test.cmd' -ea stop } | Should -Throw
@@ -186,15 +183,13 @@ Describe 'shim' -Tag 'Scoop' {
         shim-test | Should -Be 'Hello, world!'
     }
 
-    Context 'user with quote' {
-        It 'shims a file with quote in path' -Skip:$isUnix {
-            { Get-Command 'shim-test' -ea stop } | Should -Throw
-            { shim-test } | Should -Throw
+    It 'shims a file with quote in path' {
+        { Get-Command 'shim-test' -ea stop } | Should -Throw
+        { shim-test } | Should -Throw
 
-            shim "$working_dir\user with 'quote\shim-test.ps1" $false 'shim-test'
-            { Get-Command 'shim-test' -ea stop } | Should -Not -Throw
-            shim-test | Should -Be 'Hello, world!'
-        }
+        shim "$working_dir\user with 'quote\shim-test.ps1" $false 'shim-test'
+        { Get-Command 'shim-test' -ea stop } | Should -Not -Throw
+        shim-test | Should -Be 'Hello, world!'
     }
 
     AfterEach {
@@ -202,14 +197,14 @@ Describe 'shim' -Tag 'Scoop' {
     }
 }
 
-Describe 'rm_shim' -Tag 'Scoop' {
+Describe 'rm_shim' -Tag 'Scoop', 'Windows' {
     BeforeAll {
         $working_dir = setup_working 'shim'
         $shimdir = shimdir
         $(ensure_in_path $shimdir) | Out-Null
     }
 
-    It 'removes shim from path' -Skip:$isUnix {
+    It 'removes shim from path' {
         shim "$working_dir\shim-test.ps1" $false 'shim-test'
 
         rm_shim 'shim-test' $shimdir
@@ -221,7 +216,7 @@ Describe 'rm_shim' -Tag 'Scoop' {
     }
 }
 
-Describe 'get_app_name_from_shim' -Tag 'Scoop' {
+Describe 'get_app_name_from_shim' -Tag 'Scoop', 'Windows' {
     BeforeAll {
         $working_dir = setup_working 'shim'
         $shimdir = shimdir
@@ -229,11 +224,11 @@ Describe 'get_app_name_from_shim' -Tag 'Scoop' {
         Mock appsdir { $working_dir }
     }
 
-    It 'returns empty string if file does not exist' -Skip:$isUnix {
+    It 'returns empty string if file does not exist' {
         get_app_name_from_shim 'non-existent-file' | Should -Be ''
     }
 
-    It 'returns app name if file exists and is a shim to an app' -Skip:$isUnix {
+    It 'returns app name if file exists and is a shim to an app' {
         ensure "$working_dir/mockapp/current/"
         Write-Output '' | Out-File "$working_dir/mockapp/current/mockapp1.ps1"
         shim "$working_dir/mockapp/current/mockapp1.ps1" $false 'shim-test1'
@@ -246,7 +241,7 @@ Describe 'get_app_name_from_shim' -Tag 'Scoop' {
         get_app_name_from_shim "$shim_path2" | Should -Be 'mockapp'
     }
 
-    It 'returns empty string if file exists and is not a shim' -Skip:$isUnix {
+    It 'returns empty string if file exists and is not a shim' {
         Write-Output 'lorem ipsum' | Out-File -Encoding ascii "$working_dir/mock-shim.ps1"
         get_app_name_from_shim "$working_dir/mock-shim.ps1" | Should -Be ''
     }
@@ -263,35 +258,33 @@ Describe 'get_app_name_from_shim' -Tag 'Scoop' {
     }
 }
 
-Describe 'ensure_robocopy_in_path' -Tag 'Scoop' {
-    $shimdir = shimdir $false
-    Mock versiondir { $repo_dir }
-
-    Context 'robocopy is not in path' {
-        It 'shims robocopy when not on path' -Skip:$isUnix {
-            Mock Test-CommandAvailable { $false }
-            Test-CommandAvailable robocopy | Should -Be $false
-
-            ensure_robocopy_in_path
-
-            # "$shimdir/robocopy.ps1" | should -exist
-            "$shimdir/robocopy.exe" | Should -Exist
-
-            # clean up
-            rm_shim robocopy $(shimdir $false) | Out-Null
-        }
+Describe 'ensure_robocopy_in_path' -Tag 'Scoop', 'Windows' {
+    BeforeAll {
+        $shimdir = shimdir $false
+        Mock versiondir { "$PSScriptRoot\.." }
     }
 
-    Context 'robocopy is in path' {
-        It 'does not shim robocopy when it is in path' -Skip:$isUnix {
-            Mock Test-CommandAvailable { $true }
-            Test-CommandAvailable robocopy | Should -Be $true
+    It 'shims robocopy when not on path' {
+        Mock Test-CommandAvailable { $false }
+        Test-CommandAvailable robocopy | Should -Be $false
 
-            ensure_robocopy_in_path
+        ensure_robocopy_in_path
 
-            # "$shimdir/robocopy.ps1" | should -not -exist
-            "$shimdir/robocopy.exe" | Should -Not -Exist
-        }
+        # "$shimdir/robocopy.ps1" | should -exist
+        "$shimdir/robocopy.exe" | Should -Exist
+
+        # clean up
+        rm_shim robocopy $(shimdir $false) | Out-Null
+    }
+
+    It 'does not shim robocopy when it is in path' {
+        Mock Test-CommandAvailable { $true }
+        Test-CommandAvailable robocopy | Should -Be $true
+
+        ensure_robocopy_in_path
+
+        # "$shimdir/robocopy.ps1" | should -not -exist
+        "$shimdir/robocopy.exe" | Should -Not -Exist
     }
 }