+++ test/Scoop-Install.Tests.ps1
@@ -1,10 +1,9 @@
-. "$PSScriptRoot\Scoop-TestLib.ps1"
-. "$PSScriptRoot\..\lib\core.ps1"
-. "$PSScriptRoot\..\lib\manifest.ps1"
-. "$PSScriptRoot\..\lib\install.ps1"
-. "$PSScriptRoot\..\lib\unix.ps1"
-
-$isUnix = is_unix
+BeforeAll {
+    . "$PSScriptRoot\Scoop-TestLib.ps1"
+    . "$PSScriptRoot\..\lib\core.ps1"
+    . "$PSScriptRoot\..\lib\manifest.ps1"
+    . "$PSScriptRoot\..\lib\install.ps1"
+}
 
 Describe 'appname_from_url' -Tag 'Scoop' {
     It 'should extract the correct name' {
@@ -34,8 +33,8 @@ Describe 'url_remote_filename' -Tag 'Scoop' {
     }
 }
 
-Describe 'is_in_dir' -Tag 'Scoop' {
-    It 'should work correctly' -Skip:$isUnix {
+Describe 'is_in_dir' -Tag 'Scoop', 'Windows' {
+    It 'should work correctly' {
         is_in_dir 'C:\test' 'C:\foo' | Should -BeFalse
         is_in_dir 'C:\test' 'C:\test\foo\baz.zip' | Should -BeTrue
 
@@ -44,18 +43,20 @@ Describe 'is_in_dir' -Tag 'Scoop' {
     }
 }
 
-Describe 'env add and remove path' -Tag 'Scoop' {
-    # test data
-    $manifest = @{
-        'env_add_path' = @('foo', 'bar')
+Describe 'env add and remove path' -Tag 'Scoop', 'Windows' {
+    BeforeAll {
+        # test data
+        $manifest = @{
+            'env_add_path' = @('foo', 'bar')
+        }
+        $testdir = Join-Path $PSScriptRoot 'path-test-directory'
+        $global = $false
+
+        # store the original path to prevent leakage of tests
+        $origPath = $env:PATH
     }
-    $testdir = Join-Path $PSScriptRoot 'path-test-directory'
-    $global = $false
 
-    # store the original path to prevent leakage of tests
-    $origPath = $env:PATH
-
-    It 'should concat the correct path' -Skip:$isUnix {
+    It 'should concat the correct path' {
         Mock add_first_in_path {}
         Mock remove_from_path {}
 
@@ -119,39 +120,5 @@ Describe 'persist_def' -Tag 'Scoop' {
         $source, $target = persist_def @('foo', $null)
         $source | Should -Be 'foo'
         $target | Should -Be 'foo'
-    }
-}
-
-Describe 'compute_hash' -Tag 'Scoop' {
-    BeforeAll {
-        $working_dir = setup_working 'manifest'
-    }
-
-    It 'computes MD5 correctly' {
-        compute_hash (Join-Path "$working_dir" 'invalid_wget.json') 'md5' | Should -Be 'cf229eecc201063e32b436e73b71deba'
-        compute_hash (Join-Path "$working_dir" 'wget.json') 'md5' | Should -Be '57c397fd5092cbd6a8b4df56be2551ab'
-        compute_hash (Join-Path "$working_dir" 'broken_schema.json') 'md5' | Should -Be '0427c7f4edc33d6d336db98fc160beb0'
-        compute_hash (Join-Path "$working_dir" 'broken_wget.json') 'md5' | Should -Be '30a7d4d3f64cb7a800d96c0f2ccec87f'
-    }
-
-    It 'computes SHA-1 correctly' {
-        compute_hash (Join-Path "$working_dir" 'invalid_wget.json') 'sha1' | Should -Be '33ae44df8feed86cdc8f544234029fb28280c3c5'
-        compute_hash (Join-Path "$working_dir" 'wget.json') 'sha1' | Should -Be '98bfacb887da8cd05d3a1162f89d90173294be55'
-        compute_hash (Join-Path "$working_dir" 'broken_schema.json') 'sha1' | Should -Be '6dcd64f8ce7a3ae6bbc3dc2288b7cb202dbfa3c8'
-        compute_hash (Join-Path "$working_dir" 'broken_wget.json') 'sha1' | Should -Be '60b5b1d5bcb4193d19aeab265eab0bb9b0c46c8f'
-    }
-
-    It 'computes SHA-256 correctly' {
-        compute_hash (Join-Path "$working_dir" 'invalid_wget.json') 'sha256' | Should -Be '1a92ef57c5f3cecba74015ae8e92fc3f2dbe141f9d171c3a06f98645a522d58c'
-        compute_hash (Join-Path "$working_dir" 'wget.json') 'sha256' | Should -Be '31d6d0953d4e95f0a42080acd61a8c2f92bc90cae324c0d6d2301a974c15f62f'
-        compute_hash (Join-Path "$working_dir" 'broken_schema.json') 'sha256' | Should -Be 'f3e5082e366006c317d9426e590623254cb1ce23d4f70165afed340b03ce333b'
-        compute_hash (Join-Path "$working_dir" 'broken_wget.json') 'sha256' | Should -Be 'da658987c3902658c6e754bfa6546dfd084aaa2c3ae25f1fd8aa4645bc9cae24'
-    }
-
-    It 'computes SHA-512 correctly' {
-        compute_hash (Join-Path "$working_dir" 'invalid_wget.json') 'sha512' | Should -Be '7a7b82ec17547f5ec13dc614a8cec919e897e6c344a6ce7d71205d6f1c3aed276c7b15cbc69acac8207f72417993299cef36884e1915d56758ea09efa2259870'
-        compute_hash (Join-Path "$working_dir" 'wget.json') 'sha512' | Should -Be '216ebf07bb77062b51420f0f5eb6b7a94d9623d1d41d36c833436058f41e39898f2aa48d7020711c0d8765d02b87ac2e6810f3f502636a6e6f47dc4b9aa02d17'
-        compute_hash (Join-Path "$working_dir" 'broken_schema.json') 'sha512' | Should -Be '8d3f5617517e61c33275eafea4b166f0a245ec229c40dea436173c354786bad72e4fd9d662f6ac2b9f3dd375c00815a07f10e12975eec1b12da7ba7db10f9c14'
-        compute_hash (Join-Path "$working_dir" 'broken_wget.json') 'sha512' | Should -Be '7b16a714491e91cc6daa5f90e700547fac4d62e1fcec8c4b78f5a2386e04e68a8ed68f27503ece9555904a047df8050b3f12b4f779c05b1e4d0156e6e2d8fdbb'
     }
 }